UI: Some components clean up.
