+++ src/Header.jsx
@@ -10,12 +10,9 @@ import cameraManager from '../camera-manager.js'
 import metaversefile from '../metaversefile-api.js'
 import ioManager from '../io-manager.js'
 
-import { MagicMenu } from './MagicMenu.jsx';
 import { Character } from './components/general/character';
 import { Tokens } from './tabs/tokens';
 import { Claims } from './tabs/claims';
-import { Inspector } from './Inspector.jsx';
-import { Chat } from './Chat.jsx';
 import { registerIoEventHandler, unregisterIoEventHandler } from './components/general/io-handler';
 import { AppContext } from './components/app';
 
@@ -23,9 +20,9 @@ import styles from './Header.module.css';
 
 //
 
-export default function Header ({ setSelectedApp, selectedApp, userAddress, loginMethod }) {
+export default function Header ({ userAddress, loginMethod }) {
 
-    const { state, setState } = useContext( AppContext );
+    const { state, setState, selectedApp } = useContext( AppContext );
     const localPlayer = metaversefile.useLocalPlayer();
     const _getWearActions = () => localPlayer.getActionsArray().filter(action => action.type === 'wear');
 
@@ -33,9 +30,7 @@ export default function Header ({ setSelectedApp, selectedApp, userAddress, logi
     const panelsRef = useRef();
 
     const [nfts, setNfts] = useState(null);
-    const [apps, setApps] = useState(world.appManager.getApps().slice());
     const [claims, setClaims] = useState([]);
-    const [dragging, setDragging] = useState(false);
     const [wearActions, setWearActions] = useState(_getWearActions());
 
     //
@@ -46,26 +41,7 @@ export default function Header ({ setSelectedApp, selectedApp, userAddress, logi
 
     };
 
-    const selectApp = ( app, physicsId, position ) => {
-
-        game.setMouseSelectedObject( app, physicsId, position );
-
-    };
-
     //
-
-    useEffect( () => {
-
-        const update = e => {
-
-            setApps( world.appManager.getApps().slice() );
-
-        };
-
-        world.appManager.addEventListener( 'appadd', update );
-        world.appManager.addEventListener( 'appremove', update );
-
-    }, []);
 
     useEffect( () => {
 
@@ -230,55 +206,6 @@ export default function Header ({ setSelectedApp, selectedApp, userAddress, logi
 
     }, [ state.openedPanel, selectedApp ] );
 
-    useEffect( () => {
-
-        window.addEventListener('click', e => {
-
-            const hoverObject = game.getMouseHoverObject();
-
-            if (hoverObject) {
-
-                e.preventDefault();
-                e.stopPropagation();
-
-                const physicsId = game.getMouseHoverPhysicsId();
-                const position = game.getMouseHoverPosition();
-                selectApp(hoverObject, physicsId, position);
-
-            }
-
-        });
-
-    }, [] );
-
-    useEffect( () => {
-
-        const dragchange = e => {
-
-            const {dragging} = e.data;
-            setDragging(dragging);
-
-        };
-
-        world.appManager.addEventListener('dragchange', dragchange);
-
-        const selectchange = e => {
-
-            setSelectedApp( e.data.app );
-
-        };
-
-        world.appManager.addEventListener('selectchange', selectchange);
-
-        return () => {
-
-            world.appManager.removeEventListener('dragchange', dragchange);
-            world.appManager.removeEventListener('selectchange', selectchange);
-
-        };
-
-    }, [ dragging ] );
-
     const npcManager = metaversefile.useNpcManager();
     const [npcs, setNpcs] = useState(npcManager.npcs);
 
@@ -324,10 +251,7 @@ export default function Header ({ setSelectedApp, selectedApp, userAddress, logi
 
 	return (
         <div className={styles.container} onClick={ stopPropagation } >
-            <Inspector selectedApp={selectedApp} dragging={dragging} />
-            <Chat />
             <CharacterHups localPlayer={localPlayer} npcs={npcs} />
-            <MagicMenu />
             <div className={styles.inner}>
                 <header className={classnames(styles.header, styles.subheader)}>
                     <div className={styles.row}>
