+++ src/components/editor-mode/inspector/index.jsx
@@ -1,0 +1,3 @@
+
+import { Inspector } from './Inspector';
+export { Inspector };
