+++ src/components/editor-mode/inspector/inspector.module.css
@@ -1,0 +1,19 @@
+
+.inspector {
+    position: absolute;
+    width: 300px;
+    pointer-events: none;
+    user-select: none;
+    z-index: 1;
+    visibility: hidden;
+}
+
+.inspector img {
+    position: absolute;
+    top: -203px;
+    left: 0;
+}
+
+.inspector.open {
+    visibility: visible;
+}
