+++ src/components/general/user-bar/index.jsx
@@ -1,0 +1,3 @@
+
+import { UserBar } from './UserBar';
+export { UserBar };
