--- src/components/modal.css
@@ -1,49 +1,0 @@
-body {
-  background: #eee;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.modal {
-  width: 220px;
-  background: rgb(0 0 0 / 100%);
-  border: 1px solid #000;
-  transition: 1.1s ease-out;
-  box-shadow: -2rem 2rem 2rem rgba(0, 0, 0, 0.2);
-  filter: blur(0);
-  transform: scale(1);
-  opacity: 1;
-  visibility: visible;
-  position: absolute;
-  top: 150px;
-  right: -110px;
-  transform: translate(-50%, -50%);
-}
-.modal.off {
-  opacity: 0;
-  visibility: hidden;
-  filter: blur(8px);
-  transform: scale(0.33);
-  box-shadow: 1rem 0 0 rgba(0, 0, 0, 0.2);
-}
-@supports (offset-rotation: 0deg) {
-  /* offset-rotation: 0deg;
-  offset-path: path("M 250,100 S -300,500 -700,-200"); */
-  .modal.off {
-    offset-distance: 100%;
-  }
-}
-@media (prefers-reduced-motion) {
-  .modal {
-    offset-path: none;
-  }
-}
-
-.modal .content {
-  cursor: auto;
-  padding: 0.5em;
-}
-
-#centered-toggle-button {
-  position: absolute;
-}
