+++ src/components/play-mode/chat/index.jsx
@@ -1,0 +1,3 @@
+
+import { Chat } from './Chat';
+export { Chat };
