+++ src/components/play-mode/index.jsx
@@ -4,6 +4,7 @@ import React from 'react';
 import { Minimap } from './minimap';
 import { Hotbar } from './hotbar';
 import { Infobox } from './infobox';
+import { Chat } from './chat';
 
 import styles from './play-mode.module.css';
 
@@ -18,6 +19,7 @@ export const PlayMode = () => {
             <Minimap />
             <Hotbar />
             <Infobox />
+            <Chat />
         </div>
     );
